Update CommandInputType to use float for Min and Max values

+++ addons/yat/src/types/CommandInputType.cs
@@ -1,3 +1,3 @@
 namespace YAT.Types;
 
-public record CommandInputType(string Type, long Min, long Max, bool IsArray);
+public record CommandInputType(string Type, float Min, float Max, bool IsArray);
